chore: cf page add collateral and facility card
