+++ apps/admin-panel/app/credit-facilities/partial-payment.tsx
@@ -24,8 +24,8 @@ gql`
         id
         creditFacilityId
         ...CreditFacilityTransactionsFragment
-        ...CreditFacilityOverviewFragment
-        ...CreditFacilityBasicDetailsFragment
+
+        ...CreditFacilityLayoutFragment
       }
     }
   }
