+++ apps/admin-panel/app/disbursals/[disbursal-id]/voters.tsx
@@ -11,7 +11,6 @@ import {
 
 import {
   ApprovalProcessStatus,
-  GetCreditFacilityOverviewQuery,
   GetDisbursalDetailsQuery,
   GetWithdrawalDetailsQuery,
 } from "@/lib/graphql/generated"
@@ -23,7 +22,6 @@ export const VotersCard = ({
   approvalProcess:
     | NonNullable<GetDisbursalDetailsQuery["disbursal"]>["approvalProcess"]
     | NonNullable<GetWithdrawalDetailsQuery["withdrawal"]>["approvalProcess"]
-    | NonNullable<GetCreditFacilityOverviewQuery["creditFacility"]>["approvalProcess"]
     | null
 }) => {
   if (!approvalProcess) {
