+++ apps/admin-panel/app/disbursals/create.tsx
@@ -36,8 +36,7 @@ gql`
             ...DisbursalOnFacilityPage
           }
           ...CreditFacilityTransactionsFragment
-          ...CreditFacilityBasicDetailsFragment
-          ...CreditFacilityOverviewFragment
+          ...CreditFacilityLayoutFragment
         }
       }
     }
