+++ apps/admin-panel/cypress/e2e/credit-facilities.cy.ts
@@ -97,7 +97,10 @@ describe("credit facility", () => {
     cy.get('[data-testid="create-credit-facility-button"]').should("be.visible").click()
     cy.takeScreenshot("2_open_credit_facility_form")
 
-    cy.get('[data-testid="facility-amount-input"]').type("5000")
+    cy.get('[data-testid="facility-amount-input"]').type("5000", {
+      delay: 0,
+      waitForAnimations: false,
+    })
     cy.takeScreenshot("3_enter_facility_amount")
 
     cy.get('[data-testid="create-credit-facility-submit"]').click()
@@ -138,13 +141,15 @@ describe("credit facility", () => {
       .then((collateralValue) => {
         const numericValue = parseFloat(collateralValue.split(" ")[0])
 
+
+        cy.log(`Collateral value: ${numericValue}`)
         cy.get('[data-testid="update-collateral-button"]').should("be.visible").click()
         cy.takeScreenshot("7_click_update_collateral_button")
 
         cy.get('[data-testid="new-collateral-input"]')
           .should("be.visible")
           .clear()
-          .type(numericValue.toString())
+          .type(numericValue.toString(), { delay: 0, waitForAnimations: false })
         cy.takeScreenshot("8_enter_new_collateral_value")
 
         cy.get('[data-testid="proceed-to-confirm-button"]').should("be.visible")
@@ -194,7 +199,7 @@ describe("credit facility", () => {
     cy.takeScreenshot("12_click_initiate_disbursal_button")
 
     cy.get('[data-testid="disbursal-amount-input"]')
-      .type("1000")
+      .type("1000", { delay: 0, waitForAnimations: false })
       .should("have.value", "1,000")
     cy.takeScreenshot("13_enter_disbursal_amount")
