+++ apps/admin-panel/lib/apollo-client/client.tsx
@@ -11,7 +11,7 @@ import {
 } from "@/lib/graphql/generated"
 
 import { CENTS_PER_USD, SATS_PER_BTC } from "@/lib/utils"
-import { calculateBaseAmountInCents } from "@/app/credit-facilities/[credit-facility-id]/overview"
+import { calculateBaseAmountInCents } from "@/app/credit-facilities/[credit-facility-id]/collateral-card"
 
 export const makeClient = ({ coreAdminGqlUrl }: { coreAdminGqlUrl: string }) => {
   const uploadLink = createUploadLink({
